+++ src/lib/forms-schema/elements/EmailElement.ts
@@ -9,6 +9,7 @@ import {
   conditionallyShowSchemas,
   placeholderValue,
   lookupSchemas,
+  regexSchemas,
 } from '../property-schemas'
 
 export default Joi.object({
@@ -24,4 +25,5 @@ export default Joi.object({
   defaultValue: Joi.string()
     .email()
     .label('Form Element - Default Email Value'),
+  ...regexSchemas,
 })