+++ src/lib/forms-schema/elements/TextElement.ts
@@ -9,6 +9,7 @@ import {
   conditionallyShowSchemas,
   placeholderValue,
   lookupSchemas,
+  regexSchemas,
 } from '../property-schemas'
 
 export default Joi.object({
@@ -43,4 +44,5 @@ export default Joi.object({
       otherwise: Joi.string(),
     })
     .label('Form Element - Default Value'),
+  ...regexSchemas,
 })