+++ src/lib/forms-schema/property-schemas.ts
@@ -204,3 +204,21 @@ export const conditionallyShowSchemas = {
 export const storageType = Joi.string()
   .label('Storage type')
   .valid(['legacy', 'public', 'private'])
+
+const regexPattern = Joi.string()
+const regexFlags = Joi.when('regexPattern', {
+  is: Joi.string().required(),
+  then: Joi.string().regex(/^[dgimsuy]+$/),
+  otherwise: Joi.strip(),
+})
+const regexMessage = Joi.string().when('regexPattern', {
+  is: Joi.string().required(),
+  then: Joi.required(),
+  otherwise: Joi.strip(),
+})
+
+export const regexSchemas = {
+  regexPattern,
+  regexFlags,
+  regexMessage,
+}